Stop passing an ActivityControllerView as a delegate
Prior to this change, WorkoutManager had a ActivityControllerView as a
delegate. However ActivityControllerView is a struct, and therefore setting
it as a delegate will copy the struct by value, leading to strange and
unexpected behavior.

Switch to using a new class ActivitySampleProcessor. The instance of the
class can be shared between ActivityControllerView and the WorkoutManager.

Migrating variables into the class will cause some things to fail to
update; I know very little about SwiftUI so I don't know the best way to
fix these. But this at least correctly stops copying the struct.